Switch from Appraisal to Roadshow, bump versions, add 5.1
The Appraisal build was becoming really annoying to maintain, since
there's no one version of Ruby that can run the specs across all
supported versions of Rails. Roadshow is a tool that's similar to
Appraisal but uses Docker to test each version in an isolated
environment with the appropriate version of Ruby, at the cost of having
to maintain the Gemfiles manually.