Update prompt syntax for Pry v0.13.0
Version 0.13.0 of Pry changed the way that prompts are configured and 
added. Specifically, setting the prompt via an array of two procs is now 
deprecated. This commit updates the README documentation so that 
warnings are no longer emitted when running `rails console` at the 
command line and fixes #118.

See the Pry CHANGELOG for more information:
https://github.com/pry/pry/blob/master/CHANGELOG.md